Test to compute checksums
